Fix bug on click event bubbling
The original implementation didn't check the procedence of the click
event, as a consecuence, clicking on any element inside the <tr> would
delete the book.
The intended behavior is to only delete the books when the delete button
is pressed.

This was achieved with an if block to filter the events that didn't target
an element with 'submit' type.